Update grep.py
Added '-t' argument for getting number of files (and total size when used with '-s option).
Also made it so that you can choose to return the results from process_grep_query as a Python object rather than printing them. This can be useful if for example you want to compute the sizes and number of files inside all the subdirectories of a directory (which would be inconvenient to do by printing alone).